use double instead of float

+++ ipoint.c
@@ -4,11 +4,11 @@
 void reset_ipoint(struct ipoint *i)
 {
 	i->flags = 0;
-	i->k = FLT_MAX;
+	i->k = DBL_MAX;
 }
 
 
-void set_ipoint(struct ipoint *i, const struct shape *shp, int flg, float k)
+void set_ipoint(struct ipoint *i, const struct shape *shp, int flg, double k)
 {
 	i->shape = shp;
 	i->flags = FLAG_DEFINED|flg;
