+++ pinhole.h
@@ -3,7 +3,7 @@
 
 #include "types.h"
 
-bool init_pinhole(int, int, float);
+bool init_pinhole(int, int, double);
 const struct coord_system *pinhole_coord_system(void);
 
 #endif /* PINHOLE_H */
