+++ scene.h
@@ -3,8 +3,8 @@
 
 #include "types.h"
 
-extern void set_location(float, float, float);
-extern void set_target(float, float, float);
+extern void set_location(double, double, double);
+extern void set_target(double, double, double);
 extern bool intersect(struct ipoint*, const struct ray*);
 extern void add_shape(struct shape*);
 extern void prepare_shape_matrices(const struct coord_system*);
